Eth Subscription: Remove unreachable code
Removed an impossible error condition check and the associated log call.
sync.Closer can only ever return nil or "already closed", while the
removed code checked for a third error and tried to log that.

I think this was caused in an attempt to please the pedantic linter.
Issue #211 addresses the trouble with using the sync.Closer, as its
intended usage clashes with go's idioms and conventions. In retrospect,
I think I should have anticipated this problem and designed the Closer
differently from the beginning.

Signed-off-by: Steffen Rattay <steffen@perun.network>